requirements.txt
PyQt6>=6.4.0
pyqtgraph>=0.13.3
numpy>=1.21.0
scipy>=1.7.0
pyserial>=3.5 